Added basic HUD and Weapon system
WEAPONS:
Only the players has access to different weapons, I didn't see the need for this kind of system to be implemented to NPCs.
For now, you can switch weapons with up and down arrows.
For now, weapons don't do absolutely nothing, except for hud visuals, I may keep it this way in the future and make the input handle the different attacks.

HUD:
The GameHUD currently shows a health bar, handled by setting a health_actor variable; and a weapon slot, handled by using it's function.
Obviously the health bar is going to be changed, shut up.